requirements.txt
absl-py
pyparsing
numpy==1.19.5
pillow
termcolor
imageio
imageio-ffmpeg
opencv-python
xmltodict
tqdm
einops
kornia
captum
tensorboard
torchshow
scipy
imutils
torchmetrics
wandb
gym==0.20.0
timm
augmentations
h5py
ml_collections
moviepy
pandas